chore: correct default values and types
