+++ example/src/hooks/useAppContext.tsx
@@ -8,15 +8,16 @@ import {
   useCallback,
 } from 'react';
 import AsyncStorage from '@react-native-async-storage/async-storage';
-import type { PaymentConfiguration, PaymentResponse } from '../api/types';
+import type { AppConfiguration } from '../settings/types';
+import type { PaymentResponse } from '../api/types';
 import type { NavigationContainerRef } from '@react-navigation/native';
 import type { AdyenComponent } from '@adyen/react-native';
 import { isSuccess } from '../components/utilities/isSuccess';
 import { RootStackParamList } from '../router/RootStackNavigator';
 
 type AppContextType = {
-  configuration: PaymentConfiguration;
-  save: (config: PaymentConfiguration) => void;
+  configuration: AppConfiguration;
+  save: (config: AppConfiguration) => void;
   processResult: (
     result: PaymentResponse,
     nativeComponent: AdyenComponent
@@ -38,7 +39,7 @@ export const useAppContext = () => {
 const storeKey = '@config_storage';
 
 type AppContextProp = {
-  configuration: PaymentConfiguration;
+  configuration: AppConfiguration;
   onError: (error: Error) => void;
   navigationRef: NavigationContainerRef<RootStackParamList>;
 };
